refactor profiles navbar & views
