+++ eahub/templates/components/navbar.html
@@ -1,100 +1,96 @@
-{% load static %}
+{% load static sekizai_tags %}
 
 <nav class="navbar fixed-top bg-secondary navbar-dark navbar-expand-lg">
-  <div class="container">
-    {% if page_name in 'Profiles,Groups,Speakers,Candidates,Volunteers' %}
-      <div class="navbar-brand">
-        <a href="/" class="navbar__logo">
-          <img class="logo_with_page_name" id="logo" src="{% static 'global/images/logo_transparent.png' %}" />
+    <div class="container">
+        <a href="/" class="navbar-brand">
+            <img class="logo" id="logo" src="{% static 'global/images/logo_transparent.png' %}"/>
         </a>
-        <div class="navbar__page-name"> / {{ page_name }}</div>
-      </div>
-    {% elif page_name == 'Home' %}
-      <a href="/" class="navbar-brand"><img class="logo_full" id="logo" src="/static/global/images/logo_full_transparent.png" /></a>
-    {% else %}
-      <a href="/" class="navbar-brand"><img class="logo" id="logo" src="/static/global/images/logo_transparent.png" /></a>
-    {% endif %}
 
-    <div class="input-group navbar__search-input d-lg-none" data-toggle="modal" data-target="#search-box-modal">
-      <input placeholder="Search..." class="form-control" disabled>
-    </div>
+        {% block navabr_additional %}{% endblock %}
 
-
-    <button class="navbar-toggler"
+        <button class="navbar-toggler"
             type="button"
             data-toggle="collapse"
             data-target="#navbar-collapse-content"
             aria-controls="navbar-collapse-content"
             aria-expanded="false"
             aria-label="Toggle navigation"
-    >
-      <span class="navbar-toggler-icon"></span>
-    </button>
-
-    <div class="navbar-collapse collapse" id="navbar-collapse-content">
-      <ul class="navbar-nav ml-auto">
-        {% if user.is_authenticated %}
-          <li class="nav-item">
-            <a class="nav-link" href="{% url 'my_profile' %}">
-              <i class="fa fa-user"></i> My Profile 
-            </a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="{% url 'groups' %}">
-              <i class="fa fa-globe"></i> Groups
-            </a>
-          </li>
-        {% else %}
-          <li class="nav-item">
-            <a class="nav-link" href="{% url 'account_login' %}?next={% firstof request.path '/' %}">
-              <i class="fa fa-sign-in-alt"></i> Log In
-            </a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="{% url 'account_signup' %}">
-              <i class="fa fa-user-plus"></i> Sign Up
-            </a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="{% url 'groups' %}">
-              <i class="fa fa-globe"></i> Groups
-            </a>
-          </li>
-        {% endif %}
-
-        <li class="nav-item dropdown">
-          <button class="navbar-toggler nav-link d-lg-flex"
-                  type="button"
-                  aria-expanded="false"
-                  data-toggle="dropdown"
-                  aria-haspopup="true"
-                  role="button"
-                  id="navbar-collapse-lg-content"
-          >
+        >
             <span class="navbar-toggler-icon"></span>
-          </button>
-          <div class="dropdown-menu" aria-labelledby="navbar-collapse-lg-content">
-            <a class="dropdown-item" href="https://resources.eahub.org/">
-              <i class="fa fa-book"></i> EA Resources
-            </a>
-            <a class="dropdown-item" href="https://donationswap.eahub.org">
-              <i class="fa fa-exchange-alt"></i> EA Donation Swap
-            </a>
-            <a class="dropdown-item" href="{% url 'about' %}">
-              <i class="fa fa-info"></i> About
-            </a>
-            <a class="dropdown-item" href="https://feedback.eahub.org">
-              <i class="fa fa-comment"></i> Feedback
-            </a>
-            {% if user.is_authenticated %}
-              <a class="dropdown-item" href="{% url 'account_logout' %}?next={% firstof request.path '/' %}">
-                <i class="fa fa-sign-out-alt" id="logout"></i> Log out
-              </a>
-            {% endif %}
-          </div>
-        </li>
-      </ul>
-    </div>
+        </button>
 
-  </div>
+        <div class="navbar-collapse collapse" id="navbar-collapse-content">
+            <ul class="navbar-nav ml-auto">
+                {% if user.is_authenticated %}
+                    <li class="nav-item">
+                        <a class="nav-link" href="{% url 'my_profile' %}">
+                            <i class="fa fa-user"></i> My Profile
+                        </a>
+                    </li>
+                {% else %}
+                    <li class="nav-item">
+                        <a class="nav-link"
+                            href="{% url 'account_login' %}?next={% firstof request.path '/' %}">
+                            <i class="fa fa-sign-in-alt"></i> Log In
+                        </a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="nav-link" href="{% url 'account_signup' %}">
+                            <i class="fa fa-user-plus"></i> Sign Up
+                        </a>
+                    </li>
+                {% endif %}
+
+                <li class="nav-item">
+                    <a class="nav-link" href="{% url 'profiles' %}">
+                        <i class="fa fa-users"></i> Profiles
+                    </a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="{% url 'groups' %}">
+                        <i class="fa fa-globe"></i> Groups
+                    </a>
+                </li>
+
+                <li class="nav-item dropdown">
+                    <button class="navbar-toggler nav-link d-lg-flex"
+                        type="button"
+                        aria-expanded="false"
+                        data-toggle="dropdown"
+                        aria-haspopup="true"
+                        role="button"
+                        id="navbar-collapse-lg-content"
+                    >
+                        <span class="navbar-toggler-icon"></span>
+                    </button>
+                    <div class="dropdown-menu"
+                        aria-labelledby="navbar-collapse-lg-content">
+                        <a class="dropdown-item"
+                            href="https://resources.eahub.org/">
+                            <i class="fa fa-book"></i> EA Resources
+                        </a>
+                        <a class="dropdown-item"
+                            href="https://donationswap.eahub.org">
+                            <i class="fa fa-exchange-alt"></i> EA Donation Swap
+                        </a>
+                        <a class="dropdown-item" href="{% url 'about' %}">
+                            <i class="fa fa-info"></i> About
+                        </a>
+                        <a class="dropdown-item"
+                            href="https://feedback.eahub.org">
+                            <i class="fa fa-comment"></i> Feedback
+                        </a>
+                        {% if user.is_authenticated %}
+                            <a class="dropdown-item"
+                                href="{% url 'account_logout' %}?next={% firstof request.path '/' %}">
+                                <i class="fa fa-sign-out-alt" id="logout"></i>
+                                Log out
+                            </a>
+                        {% endif %}
+                    </div>
+                </li>
+            </ul>
+        </div>
+
+    </div>
 </nav>
