+++ eahub/templates/components/navbar_search.html
@@ -1,0 +1,13 @@
+{% extends 'components/navbar.html' %}
+
+
+{% load static sekizai_tags %}
+
+
+{% block navabr_additional %}
+    <div class="input-group navbar__search-input d-lg-none"
+        data-toggle="modal" data-target="#search-box-modal"
+    >
+        <input placeholder="Search..." class="form-control" disabled>
+    </div>
+{% endblock %}
