+++ eahub/templates/eahub/profiles.html
@@ -3,7 +3,7 @@
 {% load static sekizai_tags thumbnail %}
 
 {% block navbar %}
-  {% include 'components/navbar.html' %}
+  {% include 'components/navbar_search.html' %}
 {% endblock navbar %}
 
 {% block content %}
